pyproject.toml
[tool.pdm]
[tool.pdm.dev-dependencies]
dev = [
    "grpcio-tools>=1.54.0",
]
[tool.pdm.scripts]
compile_proto = "python -m grpc_tools.protoc -I./src/protos --python_out=./src/ --pyi_out=./src/ --grpc_python_out=./src/ ./src/protos/jupyter_cpp_wrapperkernel/vin/vin.proto"
docker_dev = {shell = "docker run --privileged -p=8888:8888 --rm $(docker build -q .)"}

[project]
name = "jupyter_cpp_wrapperkernel"
version = "0.0.1"
description = ""
authors = [
    {name = "Alexander Zach", email = "alexander.zach1@gmail.com"},
]
dependencies = [
    "ipykernel>=6.22.0",
    "jupyter-client>=8.2.0",
    "grpcio>=1.54.0",
    "google-api-python-client>=2.86.0",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[project.scripts]
install_cpp_wrapperkernel = "jupyter_cpp_wrapperkernel.scripts.install:main"

[project.optional-dependencies]
jupyter_notebook = [
    "notebook>=6.5.4",
]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
